Extract `darkTheme` to `theme.ts`
